fix[h265d]: Adjust condition of scan type judgement
Refer to commit 3198035973ed76fdff8c4871ddd6e20aa539dbe9.

When decoding interlaced H.265 stream, SPS will be changed after parsing
SPS NALU. This leads to comparing the new SPS and old SPS via memcpy
reporting that they are different. And references list will unreference
frames. Finally, some frame that should be kept in dpb will be removed.
And the following picture to be decoded will not find these removed
frame for reference.

Note that comparing two SPS buffer with memcpy to find if reference list
should be update is too rough. It should be fixed in the future.

Change-Id: If20ab30973df12635976d4b9d759b36b560ddea2
Signed-off-by: Johnson Ding <johnson.ding@rock-chips.com>